refactoring board-provider, moving to proper hooks
